Encourage browser chrome to disappear
Increase the vertical size of the app to slightly more than the viewport
height, so the iOS address and navigation bars will disappear when the
user vertically scrolls at all. This is not a perfect solution, but it's
better than the browser chrome always obscuring part of app in mobile
Safari.